Golang final version has passed all test.

+++ Golang-version/multi-microservices-final-version/core/src/main/java/edu/bridge/domain/UserInfo.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class UserInfo {
     private String username;
-    private Integer phoneNumber;
+    private Long phoneNumber;
     private String email;
     private String photo;
     private Float longitude;
